Flatten client, auth, and app registration into Hunter
