+++ test/hunter/report_test.exs
@@ -3,7 +3,7 @@ defmodule Hunter.ReportTest do
 
   alias Hunter.Report
 
-  @conn Hunter.Client.new(base_url: "https://mastodon.example", access_token: "123456")
+  @conn Hunter.new(base_url: "https://mastodon.example", access_token: "123456")
 
   test "reports an account with a JSON body" do
     stub_request(fn conn ->
