+++ test/hunter/result_test.exs
@@ -3,7 +3,7 @@ defmodule Hunter.ResultTest do
 
   alias Hunter.Result
 
-  @conn Hunter.Client.new(base_url: "https://mastodon.example", access_token: "123456")
+  @conn Hunter.new(base_url: "https://mastodon.example", access_token: "123456")
 
   test "searches for content with query params" do
     stub_request(fn conn ->
